fix : 통신 코드 수정 및 statistics 통신 연결
