+++ app/src/main/res/layout/fragment_show_statistic.xml
@@ -2,7 +2,7 @@
 <layout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".presentation.ui.search.ShowStatisticFragment">
+    tools:context=".presentation.ui.search.SearchStatisticFragment">
 
     <data>
         <variable
